Fix PCollectionView translation
