+++ runners/gearpump/src/main/java/org/apache/beam/runners/gearpump/translators/functions/DoFnFunction.java
@@ -135,7 +135,6 @@ public class DoFnFunction<InputT, OutputT> extends
       }
     }
 
-
     for (PCollectionView<?> sideInput: sideInputs) {
       for (WindowedValue<InputT> value : pushedBackValues) {
         for (BoundedWindow win: value.getWindows()) {
